feat: implement AdvancedFilters component and integrate it into InventoryGrid + add item type search
